feature: Pretty print using serde_json correctly
It is necessary to use serde_json::from_str to convert the config str
into a serde_json::Value then use serde_json::to_string_pretty to result
in a pretty printed String.

This reduces the number of dependencies by 3 removing getopts, jsonxf
and unicode-width so it might be worth it.